# API IoTSmartLink-Orion/API.yaml
openapi: 3.1.0
info:
  title: API IoTSmartLink-Orion - OpenAPI 3.1
  description: |-
    Esta es la versión 1.0.0 del desarrollo del API IoTSmartLink-Orion destinada al control y monitoreo de la temperatura y humedad de los refrigeradores o congeladores dentro de un laboratorico clínico. Esta API permite obtener y guardar las mediciones capturadas por una red de dispositivos IoT y observar las condiciones internas de estos congeladores y refrigeradores.

  version: 1.0.0
servers:
  - url: https://iotsmartlink-orion.swagger.io/api/v3
...
paths:
...
  /umbral:
    get:
      tags:
        - umbral
      summary: Obtener los Umbrales
      description: Se obtienen todos los umbrales de una sucursal de un laboratorio clínico
      operationId: obtenerUmbrales
      parameters:
        - name: api_key
          in: header
          description: ''
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Operación exitosa
          content:
            application/json:
              schema:
                type: array
                items: 
                  $ref: '#/components/schemas/ObtenerUmbral'
        '401':
          description: No autorizado
        '404':
          description: Umbrales no encontrados
        '400':
          description: Operación fallida
      security:
        - api_key: []
    post:
      tags:
        - umbral
      summary: Crear un Umbral
      description: Crear un Umbral para una sucursal de un laboratorio clínico
      operationId: crearUmbral
      parameters:
        - name: api_key
          in: header
          description: ''
          required: true
          schema:
            type: string
      requestBody:
        description: Crear un Umbral para una sucursal de un laboratorio clínico
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrearUmbral'
        required: true
      responses:
        '200':
          description: Operación exitosa
          content:
            application/json:
              schema:
                type: array
                items: 
                  $ref: '#/components/schemas/ObtenerUmbral'
        '401':
          description: No autorizado
        '503':
          description: Error en el servidor
        '400':
          description: Operación fallida
      security:
        - api_key: []
  /umbral/{idUmbral}:
    get:
      tags:
        - umbral
      summary: Obtener los Umbrales por idUmbral
      description: Se obtienen un Umbral por idUmbral de una sucursal de un laboratorio clínico
      operationId: obtenerUmbralPorId
      parameters:
        - name: api_key
          in: header
          description: ''
          required: true
          schema:
            type: string
        - name: idUmbral
          in: path
          description: IdUmbral necesario para la actualización
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Operación exitosa
          content:
            application/json:
              schema:
                type: array
                items: 
                  $ref: '#/components/schemas/ObtenerUmbral'
        '401':
          description: No autorizado
        '404':
          description: Umbral no encontrado
        '400':
          description: Operación fallida
      security:
        - api_key: []
    put:
      tags:
        - umbral
      summary: Actualizar un Umbral
      description: Actualizar un Umbral para una sucursal de un laboratorio clínico
      operationId: actualizarUmbralPorId
      parameters:
        - name: api_key
          in: header
          description: ''
          required: true
          schema:
            type: string
        - name: idUmbral
          in: path
          description: IdUmbral necesario para la actualización
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        description: Actualizar un Umbral para una sucursal de un laboratorio clínico
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActualizarUmbral'
        required: true
      responses:
        '405':
          description: Valores inválidos
        '401':
          description: No autorizado
        '404':
          description: Umbral no encontrado
        '400':
          description: Operación fallida
        '200':
          description: Operación exitosa
          content:
            application/json:
              schema:
                type: array
                items: 
                  $ref: '#/components/schemas/ObtenerUmbral'
      security:
        - api_key: []
    delete:
      tags:
        - umbral
      summary: Eliminar Umbral 
      description: Elimina un umbral de un contenedor de una sucursal laboratorio clínico
      operationId: eliminarUmbral
      parameters:
        - name: api_key
          in: header
          description: ''
          required: true
          schema:
            type: string
        - name: idUmbral
          in: path
          description: idUmbral para eliminar 
          required: true
          schema:
            type: integer
            format: int64  
      responses:
        '200':
          description: Se elimino el umbral
        '401':
          description: No autorizado
        '404':
          description: Umbral no encontrado
        '500':
          description: Error interno del servidor
        '400':
          description: Invalid pet value
      security:
        - api_key: []
...
components:
  schemas:
...
    ActualizarUmbral:
      type: object
      properties:
        idTipoMedicion:
          type: integer
          format: int32
          example: 1
        valorMinimo:
          type: number
          format: float
          example: 2.5
        valorMaximo:
          type: number
          format: float
          example: 5.5
        fecha:
          type: string
          format: date-time
      xml:
        name: actualizar_umbral
    CrearUmbral:
      type: object
      properties:
        idTipoMedicion:
          type: integer
          format: int32
          example: 1
        valorMinimo:
          type: number
          format: float
          example: 2.5
        valorMaximo:
          type: number
          format: float
          example: 5.5
        fecha:
          type: string
          format: date-time
      xml:
        name: crear_umbral
    ObtenerUmbral:
      type: object
      properties:
        idUmbral:
          type: integer
          format: int32
          example: 1
        idTipoMedicion:
          type: integer
          format: int32
          example: 1
        valorMinimo:
          type: number
          format: float
          example: 2.5
        valorMaximo:
          type: number
          format: float
          example: 5.5
        fecha_creacion:
          type: string
          format: date-time
        fecha_actualizacion:
          type: string
          format: date-time
      xml:
        name: obtener_umbral
...
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
    api_key:
      type: apiKey
      name: api_key
      in: header